Move the client agent bridge ABC to the agent mixins module.
