+++ src/scaler/client/client.py
@@ -113,7 +113,6 @@ class Client:
         self._future_manager: Optional[ClientFutureManager] = None
         self._bridge: Optional[ClientAgentBridge] = None
         self._connector_agent: Optional[SyncConnector] = None
-        self._connector_storage: Optional[SyncObjectStorageConnector] = None
 
         self._serializer = serializer
 
@@ -153,18 +152,14 @@ class Client:
         self._connector_agent = self._bridge.connector
 
         logger.info(f"ScalerClient: connect to object storage at {self._object_storage_address}")
-        self._connector_storage = self._backend.create_sync_object_storage_connector(
-            identity=self._identity, address=self._object_storage_address, security_config=self._security_config
-        )
 
-        self._object_buffer = ObjectBuffer(
-            self._identity, self._serializer, self._connector_agent, self._connector_storage
-        )
+        # The object storage server is accessed by the client agent, on its event loop.
+        self._object_buffer = ObjectBuffer(self._identity, self._serializer, self._connector_agent, self._bridge)
         self._future_factory = functools.partial(
             ScalerFuture,
             serializer=self._serializer,
             connector_agent=self._connector_agent,
-            connector_storage=self._connector_storage,
+            object_buffer=self._object_buffer,
         )
 
     @property
@@ -796,9 +791,6 @@ class Client:
 
         if self._connector_agent is not None:
             self._connector_agent.destroy()
-
-        if self._connector_storage is not None:
-            self._connector_storage.destroy()
 
     @staticmethod
     def __get_parent_task_priority() -> Optional[int]:
